Users/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
databases[postgresql]==0.6.0
asyncpg==0.30.0
redis==4.5.5
graphene==3.3
starlette-graphene3==0.6.0
cryptography==41.0.2
python-dotenv==1.0.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.14.1
pytest==8.3.3
pytest-asyncio==0.26.0
pytest-cov==6.0.0
httpx==0.27.2
psycopg2-binary